Insert a link when know we have a URL. Fix pasting URLs on iOS.
Pasting a URL on iOS is broken since it doesn’t put a text/plain
representation of the URL on the clipboard, but only a text/uri-list.
This uses the ‘URL’ type which is a shortcut for retrieving the first
URL found. Also improves handling of URLs by inserting a link