Toolbar buttons are removed from the navigation flow
Every toolbar button gets a tabindex property with a value of -1. It removes the buttons
from the navigation flow, meaning a user cannot tab to it. It improves usability when
Trix is placed after a focusable field (i.e. the user can tab directly to the content
area). It also improves the accessibility of Trix.