Altered setAttributes to always return a promise
setAttributes sometimes delves into the `preload` method that uses a
promise (operation) to wait for the source file to load into the
browser before doing some additional things. Thus, setAttributes should
also return a promise so that third-party code using this method can
wait for it to complete before doing other things such as allowing the
content of the editor to be saved.

I’ve changed the function so that we wait for `didChangeAttributes()`
to complete before executing `attachmentDidChangeAttributes()` on the
delegate as before this was triggering a premature notification.

I’ve not added any new tests but have adjusted the only one testing
this functionality so that it waits for the promise to resolve before
asserting.