Downgrade selenium-webdriver to avoid flaky iOS tests
Hi Javan!

Based on the error message, this appears to be the issue  that i've seen here and there where if the code is using the Java Selenium 3.x bindings, Appium has issues parsing the capabilities and it produces the error that you see.

If you downgrade your bindings to 2.53.1 or use the 3.2.0 bindings, do you see any improvements?

Best,
Chip Vaughn
Customer Support Engineer