Change default height of editor to 6.25em
I have been using Trix for a while now and I found it very weird that styling the editor was something I always had to do. I have been wondering if we could move the default min-height of the editor to 6.25em which is somewhat round and moderately long enough for a text-area.