Add nested `code` tags to `pre` tags in code block
Have added the a new `pre` block attribute for the `pre` tag and added
the `code` tags as a "list item" within the `pre` tags. Adding the
nested `code` tag should make it possible to style codeblocks with
syntax highlighters going forwards.

Many of the tests for the `code` tag had to be updated to take this
change into account. The test below referred to an incorrect Document
Range and so had to be changed significantly:

  - Block formatting: backspace selection spanning and entire formatted
    block and a formatted block