Ensure trix-editor element is ARIA labeled
Since the `<trix-editor>` element is custom and not a browser `<input>`,
we can't rely that assistive technology will be able to [read the label
text from a `<label for="...">` that corresponds to a `<trix-editor
id="...">` element][labelable].

When a `<trix-editor>` element is attached to the DOM and initialized,
collect the text from its related `<label>` elements and combine it to
set a default `[aria-label]` value.

If the `<trix-editor>` element is a descendant of any of its `<label>`
elements, *omit* those elements when generating an `[aria-label]` value.

If an [`[aria-label]` value][aria-label] is already specified, *do not*
override it.

Similarly, if a [`[aria-labelledby]` value][aria-labelledby] is
specified, don't set the `[aria-label]` attribute at all.

[labelable]: https://html.spec.whatwg.org/multipage/forms.html#category-label
[aria-label]: https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/ARIA_Techniques/Using_the_aria-label_attribute
[aria-labelledby]: https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/ARIA_Techniques/Using_the_aria-labelledby_attribute