Allow intercepting attachment toolbar creation
There are cases where applications want to customize the attachment
toolbar UI to add additional functionality. To allow this, we add a new
`trix-attachment-before-toolbar` event which fires when an attachment is
selected in the editor, before its toolbar is displayed. Applications
can listen for this event, and alter or extend the toolbar as necessary.

The event contains the follow attributes:

- `attachment` - the `Attachment` instance corresponding to the item
  that is being selected.
- `toolbar` - the element containing the toolbar UI. Applications can
  modify this element and its children as necessary.