Remove check for DOM range equality in SelectionChangeObserver
This checks causes problems on Safari 17 when you drag the cursor to
select text. Safari weirdly reports the selection range as being
collapsed, even when you select multiple characters. This causes the
SelectionManager to think that the selection is collapsed and not
trying to find where it ends. The end result for the user is that
they can not apply formatting or links to the whole selected text.

This is ultimately a Safari bug, but it's not clear that we need to
check for DOM range equality in the first place. The only reason we
do this is because the original implementation of the
SelectionChangeObserver used polling to check for selection changes,
instead of relying on the `selectionchange` event. Now that we're
using the `selectionchange` event, we can assume that if a selection
change event fires, the selection has changed.