Use the target range to insert the replacement text
When the browser auto-corrects a misspelled word, it dispatches a
`beforeinput` event with the `insertReplacementText` input type and a
`dataTransfer` object containing the replacement text. The event also
contains a `getTargetRanges` method that returns an array of `DOMRange`
objects representing the range of text that will be replaced.

When the `Level2InputController` was originally implemented, not all
browsers supported the `getTargetRanges` method, so it was not used.

Now that all supported browsers do support `getTargetRanges`, we can use
it to insert the replacement text at the correct location. This ensures
editor state is updated correctly the cursor is positioned correctly.

Ref.

- https://w3c.github.io/input-events/#overview (search for "insertReplacementText")
- https://developer.mozilla.org/en-US/docs/Web/API/InputEvent/getTargetRanges