`ElementInternals`: Support without `<input>` element
The successful migration to utilize [Element Internals][] stopped short
of removing Trix's dependency on an associated `<input type="hidden">`
element to store its state. Prior to integration with
`ElementInternals`, Trix relied on the `<input>` for the sake of having
its value serialized into a form submission. That is no longer
necessary.

When Trix is configured to be compatible with `ElementInternals`, it is
also capable of functioning without an `<input type="hidden">` element.
To configure a `<trix-editor>` element to skip creating its `<input
type="hidden">`, set the element's `willCreateInput = false`:

```js
addEventListener("before-trix-initialize", (event) => {
  const trixEditor = event.target

  trixEditor.willCreateInput = false
})
```

Trix will *always* use an associated `<input type="hidden">` element
when the `[input]` attribute is set, regardless of its `willCreateInput`
property. To migrate to `<input>`-free support, render the
`<trix-editor>` without the `[input]` attribute.

In the absence of an `<input type="hidden">` element, the
`<trix-editor>` element's value will not be included in `<form>` element
submissions unless it is rendered with a `[name]` attribute. Set the
`[name]` attribute to the same value that the `<input type="hidden">`
element would have.

Additional considerations
---

Action Text integration should remain unchanged. At the time of this
submission, Rails automatically renders an `<input type="hidden">`
element that is paired with the `<trix-editor>` element through its
`[input]` attribute.

[Element Internals]: https://developer.mozilla.org/en-US/docs/Web/API/ElementInternals