Prioritize pasting files over pasting HTML content
When right clicking and pasting an image the dataTransfer object
contains a file and text/html. If we paste the HTML content, the result
is often a broken image. This change prioritizes pasting the file over
pasting the HTML content.

The gotcha is that when pasting text from MS Word, the dataTransfer object
contains a File with a screenshot of the text and text/html with the
actual text. We need to check if the paste is from MS Word and if so,
we should paste the text instead of the screenshot.