Prevent duplicate file pastes
This change prevents duplicate file pastes by checking if the paste event
is a file paste before handling it. This is necessary because Safari
doesn't support `beforeinput.insertFromPaste` for files, so we are handling
file pastes in the paste event handler too.

We need to check in the `beforeinput.insertFromPaste` event handler if the
event will be also handled by the paste event handler, to avoid duplicate
file pastes.