Prioritize pasting files over pasting HTML content (#1148)
* Prioritize pasting files over pasting HTML content

When right clicking and pasting an image the dataTransfer object
contains a file and text/html. If we paste the HTML content, the result
is often a broken image. This change prioritizes pasting the file over
pasting the HTML content.

The gotcha is that when pasting text from MS Word, the dataTransfer object
contains a File with a screenshot of the text and text/html with the
actual text. We need to check if the paste is from MS Word and if so,
we should paste the text instead of the screenshot.

* Prevent duplicate file pastes

This change prevents duplicate file pastes by checking if the paste event
is a file paste before handling it. This is necessary because Safari
doesn't support `beforeinput.insertFromPaste` for files, so we are handling
file pastes in the paste event handler too.

We need to check in the `beforeinput.insertFromPaste` event handler if the
event will be also handled by the paste event handler, to avoid duplicate
file pastes.