Document the `trix-action-invoke` event
The [Trix Wiki][wiki] has a page documenting how to [customize the
Toolbar element][toolbar]. Its a helpful, thorough document with a lot
of great information, including how to invoke custom actions through the
rendering elements with an `x`-prefixed `[data-trix-action]` attribute.

Since this information is at the bottom of the wiki page, it isn't
readily available from the `README.md`. The `trix-action-invoke` event
that's dispatched is also omitted from the README's list of events.

This commit promotes the contents of the wiki page into the `README`,
documents the `trix-action-invoke` event, and adds missing test coverage
that exercises the event.

[wiki]: https://github.com/basecamp/trix/wiki/Configuration
[toolbar]: https://github.com/basecamp/trix/wiki/Configuration#toolbar