Make Trix compatible with morphing
Morphing the trix editor will leave it in an unusuable state because
morphing will replace the bootstrapping elements without triggering
the connection logic.

The solution is based on adding an attribute to flag the editor as
initialized and detect when that attribute is modified.

See https://github.com/hotwired/turbo-rails/issues/533#issuecomment-1842056856