Build `action_text-trix` assets with existing tools
Removes the JavaScript and CSS file `FileUtils.cp` calls from
`action_text-trix/Rakefile` so that the files can be generated by the
existing tooling (`rollup.config.js` for JavaScript, `bin/sass-build`
for CSS).

With this change, the project is free to continue to utilize those tools
to manage the files that will be included in the Gem, without needing to
span the Ruby-NodeJS boundary.